+++ Extension/src/common/forward.ts
@@ -80,6 +80,7 @@ export const enum ForwardAction {
     LearnAboutAdGuard = 'learn_about_adguard',
     FilteringLogAssumedRule = 'filtering_log_assumed_rule',
     Christmas24 = 'christmas_24',
+    CustomFiltersMv3Disabled = 'custom_mv3_disabled',
 }
 
 export const enum ForwardFrom {