+++ tests/helpers/fixtures/getDefaultExportFixture.ts
@@ -11,6 +11,7 @@ import {
 } from '../../../Extension/src/background/schema';
 import { UserAgent } from '../../../Extension/src/common/user-agent';
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const getDefaultExportFixture = (isMv3: boolean): Config => ({
     [RootOption.ProtocolVersion]: PROTOCOL_VERSION,
     [RootOption.GeneralSettings]: {
@@ -31,8 +32,10 @@ export const getDefaultExportFixture = (isMv3: boolean): Config => ({
         [ExtensionSpecificSettingsOption.UserRulesEditorWrap]: false,
     },
     [RootOption.Filters]: {
-        // For MV3 we added AdGuard Quick Fixes filter enabled by default.
-        [FiltersOption.EnabledFilters]: isMv3 ? [2, 10, 24] : [2, 10],
+        // TODO: uncomment this when quick fixes filter will be supported for MV3
+        // // For MV3 we added AdGuard Quick Fixes filter enabled by default.
+        // [FiltersOption.EnabledFilters]: isMv3 ? [2, 10, 24] : [2, 10],
+        [FiltersOption.EnabledFilters]: [2, 10],
         [FiltersOption.EnabledGroups]: [1, 6],
         [FiltersOption.CustomFilters]: [],
         [FiltersOption.UserFilter]: {